Move get_class_name into cruby
